// src/utils/enums/store.enum.ts
export enum StoreEnum {
	ADD_POST = 'ADD_POST',
    UPDATE_NEW_POST_TEXT = 'UPDATE_NEW_POST_TEXT',

    ADD_MESSAGE = 'ADD_MESSAGE',
    UPDATE_MESSAGE = 'UPDATE_MESSAGE',

    USER_SUBSCRIBE = 'USER_SUBSCRIBE',
    USER_UNSUBSCRIBE = 'USER_UNSUBSCRIBE',

    SET_USERS = 'SET_USERS',
    SET_CURRENT_PAGE = 'SET_CURRENT_PAGE',
    SET_TOTAL_USERS_COUNT = 'SET_TOTAL_USERS_COUNT',

    TOGGLE_IS_FETCHING = 'TOGGLE_IS_FETCHING'
}